busybox-mdev: start mdev in S04
initscripts-1.0 starts 'devices' in S05, and expects
udev to be running in S04. As busybox-mdev replaces
udev, it should run before the 'devices' script as well,
so we start it in S04.

Note that it makes no sense to bump PR, because update-rc.d
will not allow S04mdev to be created when S06mdev still exists,
 and there is no pkg_postrm script for mdev.
So this change requires a clean rebuild/install.