busybox-mdev: Support disks without partitions
Flash devices may be formatted without creating a partition. To recognize it,
an attempt must be made to mount '/dev/hdc' directly. If the disk has a proper
partition table, that mount will fail, and the mdev script will trigger again
when /dev/hdc1 is detected.

This fixes that some CF cards and some USB flash sticks were not being mounted,
while others were.

DM7025: Also added a small change to only send the hdparm command to the device,
now that the device itself also triggers the script.