initscripts: Unmount network filesystems before the network stops
This fixes the 7025 not shutting down at all and long timeouts on
other boxes when there are open network shares.

unmounting network filesystems "hangs" because the network is down
when the umount scripts fire.
Network stops at K40, so umountnfs should run before that.

Created a simplified version of the scripts for openpli that also
mount and umount CIFS shares.