opkg: use depmod -a in depmod interceptor
depmod -A was used (=update only when new modules are found)
With this option, depmod compares the ko timestamps with the
modules.dep timestamp, and decides whether an update is required.
This does not work in our case, when opkg installs new ko's,
their timestamp remains at the build date.
So depmod never detects changes, and does not rebuild modules.dep.
Also, when opkg calls the depmod interceptor, of course this means
that new modules were installed, so depmod -A would actually
have more overhead than depmod -a.