dm7025/enigma2: Place all the Python code into an archive
This should allow a 7025 to do an upgrade and survive. However, this
time a flash is highly recommended.

Faster upgrade, since only a single file is changed instead of dozens.
Less flash space, the archive is expected to be smaller.

Drawback is that installing enigma2-src will cause recompilation of
all .py files, wasting lots of flash and CPU time. Hence this is only
done for the 7025 for now.