dreambox-boottool-cf.bb: Boot dm7025 from CF card
Attempts to boot from CF after installing this plugin.

To make this work (I need to put this in a utility...):
- CF card must have been formatted in ext3.
- Copy a root system to CF, create the mountpoints etc.
- Reboot and hope for the best.

If it fails to boot, it will try to boot from old-fashioned
flash instead.

On disaster, just flash your dreambox again... You may need to
remove the CF card due to configuration issues (opkg will think
that the rootfs on the CF actually means something).