opkg-collateral: Remove USB/CF/HDD destinations from opkg.conf
The only reason they were in this file is to be able to install picons there. This has
been changed already, allowing picons on any mounted media without requiring its
presence in opkg.conf.

This prevents the creation of /usr/lib/opkg directories on these paths.

This also solves a problem with the 7025: After running CF boot, flashing
a new image results in an almost "dead" machine because all the configure
scripts go south on the existing /media/cf/usr/.. metadata.