openpli-image: use busybox-inetd instead of inetd
We cannot rely on PREFERRED_PROVIDER_inetd in this case,
because IMAGE_INSTALL is used in both RDEPENDS and DEPENDS,
and inetd itself is a package name.
So inetd will be built because inetd is in our image DEPENDS,
but busybox-inetd (which is built anyway) ends up in the
image because the RDEPENDS obeys the PREFERRED_PROVIDER.

Strictly speaking, the inetd provider should of course
not be decided by the image, but by the distro.
However, in this case that would mean we cannot avoid
'multiple provider' warnings without having to introduce
a virtual.
This workaround in the image recipe has less impact.