linux-etxx00: Add fuse internally, disable internal ntfs
Load fuse early (because ntfs-3g wants it).
NTFS in kernel cannot create or grow files. On the etxx00 we use
ntfs-3g to provide writable NTFS mounts anyway, so the kernel one
is not of much use (so built as module for who wants it).

Still investigating why an NTFS disk is not mounted at boot (but
mounts okay when you connect it later).