Revert "sysvinit: start getty on the console tty"
This reverts commit 9bdbd63487e92ce4560cfd92dc13213e03f65696.

A getty on a serial console does not work unless you start it on the
comport devicenode. Starting it on tty1 does not work.
And a hardcoded /dev/ttyS0 will not do, because that does not allow
the comport to be left alone with console=null

When a certain platform needs a console, we'd better configure it
with SYSVINIT_ENABLED_GETTYS in the machine.conf.