linux-etxx00: dvb-usb patches by Gennar1
dvb-usb-rtl2832.patch is an update for the Realtek 2832 driver;
it adds support for kernels >= 3.3 as well as for the current
media_build tree; also, it fixes a bug in the GetSignalStrength
function so now the signal strength should be scaled to the proper
range (it used to be capped at about 39%).

dvb-usb-it913x-patch-collection.patch contains a few updates from
the official mantainer of the driver on the linux-media mailing list;

em28xx-pre-allocate-DVB-iso-transfer-buffers.patch is completely new
and it implements pre-allocation of the USB isoc transfer buffers in
the em28xx driver (in digital mode); this avoids the crashes due to
memory fragmentation that are currently plaguing the em28xx driver
on the MIPS architecture.