djmount: correct RDEPENDS and init-file
djmount does not RDEPENDS on "fuse-utils". It only needs libfuse, which is build in package "fuse".

The third issue is, that if you want to remove the djmount-ipk,
the command "/etc/init.d/djmount stop" is fired.
If djmount is not actually active, start-stop-daemon gives an error
and the remove sequence fails.
Then opkg remove can only be done by "--force-remove"
This patch let always give back a "true" when djmount stop is fired,
so that opkg remove will not fail.

Signed-off-by: pieterg <pieterg@users.sourceforge.net>